issues: Improve comment handling
If someone attempts to add all whitespace, we should reject this, as
it isn't a proper comment and could signal that the editor merely
left a file that got a single newline or something.

Additionally, there is a missing sys import, so might as well glob
that with this one.

Signed-off-by: Aaron Conole <aconole@redhat.com>