issues: Allow a '-' to indicate no assignee.
Having the issues list always populate assignee is quite troublesome, because there
may be a reason for a user to query other issues that have interesting flags set
but which aren't assigned to that user.  In order to do this, we want a special
flag that allows the issue list to override the default.  Choose '-' but if there
is some objection, we can change it.

Signed-off-by: Aaron Conole <aconole@redhat.com>