boards: Fix up the assignee filter.
We do need to filter in the cli command because the board jql is fixed,
and rather than trying to rewrite it programmatically, we will post-process
the issues.  When doing this filtering, it is important to check that
there is an actual assignee before matching on fields in the assignee
object.

Signed-off-by: Aaron Conole <aconole@redhat.com>