boards: Add a get-config to dump board details.
It can be useful to see the specific JQL used to run against a board
to help determine which issues will be selected.  Add the config get
to dump the column details, and filter.  A future enhancement would
be to dump the admins and other details, but that isn't as useful.

Signed-off-by: Aaron Conole <aconole@redhat.com>