boards: Add support for the quickfilter extensions.
JIRA Boards can allow quickfilters to be defined that extend the core
filter for a board by allowing subfilters (such as on a project or
particular user).  It is conveneient to import such functionality
because when selecting issues and trying to post-process, we may
run into limitations with the max number of issues allowed, or tracking
starting offsets.

Suggested-by: Eelco Chaudron <echaudro@redhat.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>