issues: Add a creation routine for creating issues.
When working with JIRA, occasionally a developer comes across an issue
in code while working on something else.  This can sometimes require
creating a JIRA ticket, but doing that can sometimes break the workflow
that the developer uses in the CLI.  Having to fight with a browser's
auth fixtures or find the right tabs, etc.

In order to combat this, add a command to create an issue and make
the issue creation similar to adding a commit via the git-commit
workflow.  The issue creation interface uses a simple text editor
and recognizes comments (lines that start with a #) to be excluded from
the issue creation, but used for controlling fields.  As an example,
we can set issue-type, project, and custom fields to the requested
values by using the appropriate comment-block control.

Additionally, the interface supports parsing a git patchfile and
populating description and summary types automatically.  This can
be useful for working with upstream patches that are submitted.
Also, it can cover the cases where a developer finds a problem in
a particular bit of code and wants to quickly attach a JIRA issue
to the associated project.

Signed-off-by: Aaron Conole <aconole@redhat.com>