issues: Include a way to get all the fields for setting.
When creating an issue, it can be difficult to determine which fields are okay
to set at the beginning.  JIRA is very picky and not all projects will include
all fields.  Add a convenience option to show all the fields of major concern
in a project.  Other fields may be settable later, but these are the fields
that could be set at issue creation time.

Signed-off-by: Aaron Conole <aconole@redhat.com>