issues: Allow setting fields from the command line.
Filling in the set-fields options for a file will work, but we also can
support setting fields from the command line.  While the option was there
it didn't actually do anything.  Implement it.  A future enhancment would
be to make it merge with the assignable fields from the bug itself so we
skip adding those.

Signed-off-by: Aaron Conole <aconole@redhat.com>