issues: Add sorting option
This should allow for sorting the issues list to make the important stuff
bubble to the top or bottom of the view.  Alter the query builder to recognize
a special field called "ORDER BY" which will be passed over and then
passed to the internal orderby code.  The orderby builder also can be passed
an already assembled "ORDER BY" string and it will happily accept it.

Signed-off-by: Aaron Conole <aconole@redhat.com>
Suggested-by: Flavio Leitner <fbl@redhat.com>