connector: Allow 'array' to push objects
Previously, all array types were treated as simple types and sent
as-is.  This meant that we rely on the jira server to do the conversion
to the proper internal representation.  Some fields, however, are complex
options and require objects with proper specifiers.  Spelling those out
is difficult, and the pyjira API doesn't include a good way of figuring
out what can be set at the moment.

For now, allow a user to give a complex type as a pydict object.  This
can be passed to the jira server and will be converted properly to the
object, provided the spelling matches.

Reported-by: Marcelo Leitner <mleitner@redhat.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>