issues: Allow forcing a value.
Sometimes the JIRA advertised schema isn't actually reflective of what
needs to be set.  One way of dealing with this is to try and parse the
error conditions and re-submit the request after massaging the data.
That is very error prone, and too complicated to implement for a scripting
interface.  However, we can give some power to the user by allowing
them to set the field directly with a 'force' flag that will do an
eval and just use the resulting value.

Signed-off-by: Aaron Conole <aconole@redhat.com>