comments: Add a reply-to method.
Adrian Moreno noted that it would be very handy to have a method
for replying to comments that includes a pre-formatted reply
post.  The current method involves multiple windows and copying
data between them.  With this patch, we introduce a new option
'--in-reply-to' that allows users to get the contents of the
existing comment for editing.

Also folded in a patch to the status field that makes them all
standard size.  This prevents random failures.

Suggested-by: Adrian Moreno <amorenoz@redhat.com>
Signed-off-by: Adrian Moreno <amorenoz@redhat.com>
Co-Authored-by: Adrian Moreno <amorenoz@redhat.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>