markdown: support multiline quotes.
Markdown quotes can be empty lines, i.e:

>·One·line
>
>·Another

(Artificially replaced " " with "·")

However, when automatically generating quotes (as in --reply-to), the
easy (and clean) way is to append "> " to every line, just as email
clients do, resulting in:

>·One·line
>·
>·Another

(Artificially replaced " " with "·")

Both blocks should be translated into:

{noformat}
One·line

Another
{noformat}

This patch attemps to support that.
Also, test strings are made multi-line strings to reduce eye-strain.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>