issues: fix visibility inheritance in replies
Current visibility inheritance is based on the option being None. But
that actually never happens betcause it has a default value of "all".

Use the click's ParameterSource to determine if "all" came from the
default (in which case, inheritance should override it) or was
explicitly set by the user (in which case, visibility inheritance is
skipped).

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>