issues: fix the create call.
Commit 1d7e1a8f23566 ("issues create: add new line to get proper formatting")
introduced a way to preserve the jira connection object across calls.
This is quite useful.  However, it assumes an obj context is present
when that isn't always the case.  Properly guard against the missing
obj dict by ensuring it exists.

Also, introduce a test to ensure we don't miss future changes to the
create command.

Fixes: 1d7e1a8f23566 ("issues create: add new line to get proper formatting")
Acked-by: Marcelo Leitner <mleitner@redhat.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>