issues: Do not require template file to exist for all modes.
The template file should be optional, and only required for the template mode.
Unfortunately, click doesn't provide a good way to indicate this.  For now, we
will check that the argument is used and bail before even logging in.  In the
future, a better way of doing option dependencies should be investigated.

Signed-off-by: Aaron Conole <aconole@redhat.com>