boards: Add the ability to dump JSON and ignore issues.
JIRA board queries can be long running operations because we query for
all the issues in a sprint.  These queries aren't always needed, like
when we just want to nab the board id.

Additionally, if we want to nab the board id, it isn't easy because
the output is more like a human format.  So add a JSON dump as well.

Signed-off-by: Aaron Conole <aconole@redhat.com>