issues: Always output all the json fields.
Ondrej Lichtner notes that when the empty set is returned for a
list, the JSON formatting is invalid.  Looking a bit closer, it
doesn't seem to make sense to skip printing the fields for an
empty set of issues.  Change it so that we print all the fields
now, making the JSON consistent and valid.

Reported-by: Ondrej Lichtner <olichtne@redhat.com>
Reported-at: https://github.com/apconole/jiracli/issues/16
Signed-off-by: Aaron Conole <aconole@redhat.com>