issues: Fix the issue creation regex with field setters.
During f0988025f6a6 ("issues: Allow issue creation to pass the forced flag.")
the regex was modified to allow passing the forced flag.  Unfortunately, that
caused the field matching code to require --forced to be included due to the
regex missing a crucial '?' to allow for 0-1 instances of --forced to exist.

Fixes: f0988025f6a6 ("issues: Allow issue creation to pass the forced flag.")
Reported-by: Chris van Hoof <vanhoof@ouwish.com>
Reported-at: https://github.com/apconole/jiracli/issues/17
Signed-off-by: Aaron Conole <aconole@redhat.com>