auth: Allow for browser cookie harvesting.
Users who want to use jiracli have complained that acquiring
the personal access token is quite confusing.  Rather than
writing some kind of automation to help get a PAT, we use
the browser_cookies3 module to pull the cookies from an
active browser session.  This also has the advantage that
we can skip needing to store access credentials in an
on-disk file.  The downside is that users will need to
log in via the browser when the cookies expire (which
defeats the nicety of a cli interface).  The upside of
being able to start right away seems to outweigh the
downside.

Signed-off-by: Aaron Conole <aconole@redhat.com>