boards: Properly handle sprint pagination.
The JIRA client 'sprints' API will default to listing all of
the sprints associated with the board.  There are two ways to
handle this.  The first way is to pass state='active,future'
and allow the API to deal with the results.  This has some
problems after testing, notably that it isn't any faster and
we still could need to paginate results (if there are enough
future sprints on the board).  The second way is to keep doing
requests and accumulate the results, as a pagination.  This
will preserve the ability to look at closed sprints.

Reported-by: Chris van Hoof <vanhoof@ouwish.com>
Reported-by: Christian Trautman <ctrautma@redhat.com>
Signed-off-by: Aaron Conole <aconole@redhat.com>